+++ flipper-backstack/src/main/java/fr/afaucogney/mobile/flipper/internal/callback/FlipperFragmentCallback.kt
@@ -13,6 +13,7 @@ internal class FlipperFragmentCallback(
 
     override fun onFragmentAttached(fm: FragmentManager, f: Fragment, context: Context) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_ATTACHED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentCreated(
@@ -21,6 +22,7 @@ internal class FlipperFragmentCallback(
         savedInstanceState: Bundle?
     ) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_CREATED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentActivityCreated(
@@ -29,6 +31,7 @@ internal class FlipperFragmentCallback(
         savedInstanceState: Bundle?
     ) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_ACTIVITY_CREATED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentViewCreated(
@@ -38,22 +41,27 @@ internal class FlipperFragmentCallback(
         savedInstanceState: Bundle?
     ) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_VIEW_CREATED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentStarted(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_STARTED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentResumed(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_RESUMED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentPaused(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_PAUSED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentStopped(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_STOPPED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentSaveInstanceState(
@@ -62,23 +70,28 @@ internal class FlipperFragmentCallback(
         outState: Bundle
     ) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_SAVE_INSTANCE_STATE)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentViewDestroyed(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_VIEW_DESTROYED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentDestroyed(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_DESTROYED)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     override fun onFragmentDetached(fm: FragmentManager, f: Fragment) {
         flipperPlugin.pushFragmentEvent(f, FragmentLifeCycle.ON_FRAGMENT_DETACHED)
         flipperPlugin.moveToTrashAndUpdate(f)
+        flipperPlugin.pushFragmentManagerEvent(f, fm)
     }
 
     internal interface IFragmentLifeCycleCallbackFlipperHandler {
         fun pushFragmentEvent(fragment: Fragment, event: FragmentLifeCycle)
+        fun pushFragmentManagerEvent(fragment: Fragment, fm: FragmentManager)
         fun moveToTrashAndUpdate(fragment: Fragment)
     }
 }