+++ flipper-backstack/src/main/java/fr/afaucogney/mobile/flipper/internal/model/FlipperExt.kt
@@ -5,9 +5,16 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModel
 import com.facebook.flipper.core.FlipperObject
 
-const val FID = "id"
-const val NAME = "name"
+const val BACK_STACK = "backStack"
+const val ENTRIES = "entries"
+const val ADDED_FRAGMENTS = "addedFragments"
+const val ACTIVE_FRAGMENTS = "activeFragments"
+
+const val FRAGMENT_MANAGER = "fragmentManager"
 const val FULL_NAME = "fullName"
+const val FID = "id"
+const val LIFE_CYCLE_EVENT = "lifeCycle"
+const val NAME = "name"
 const val TYPE = "type"
 
 const val NEW_DATA_KEY = "newData"
@@ -16,19 +23,17 @@ const val FILTER_OPTION_KEY = "newTreeFilterOptions"
 
 const val FILTER_OBJECT_TREE_KEY = "options"
 
-const val LIFE_CYCLE_EVENT = "lifeCycle"
-const val BACK_STACK = "backStack"
-const val TRASH = "Trash"
-
+// OPTIONS
 const val APPLICATION = "Application"
 const val ACTIVITIES = "Activities"
 const val FRAGMENTS = "Fragments"
 const val VIEW_MODELS = "ViewModels"
 const val JOBS = "Jobs"
 const val SERVICES = "Services"
-const val MEMBERS = "LiveDatas"
-const val JETPACK_BACKSTACK = "JetPackBackStack"
-const val LEGACY_BACKSTACK = "LegacyBackStack"
+const val VIEW_MODEL_MEMBERS = "LiveData"
+const val BACKSTACK_JETPACK = "JetpackBackStacks"
+const val BACKSTACK_LEGACY = "LegacyBackStacks"
+const val TRASH = "Trash"
 
 const val TIMESTAMP = "timestamp"